requirements.txt
# Core web app runtime
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# Model dependencies
numpy>=1.26.0
torch>=2.3.0
transformers>=4.40.0
scikit-learn>=1.3.0

# Legacy
aiohttp>=3.9.0
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
twitchio>=2.9.0
